Created a 'common' interface for case and task
