+++ api/workAllocation2/interfaces/task.ts
@@ -35,62 +35,3 @@ export interface TaskAction {
   id: string;
   title: string;
 }
-
-export interface Assignee {
-  id: string;
-  userName: string;
-}
-
-export interface CaseData {
-  category: string;
-  location: LocationSummary;
-  name: string;
-  reference: string;
-}
-
-export interface LocationSummary {
-  id: string;
-  locationName: string;
-}
-
-export interface Location {
-  id: string;
-  locationName: string;
-  services: string[];
-}
-
-export interface LocationApi {
-  location_id: string;
-  location: string;
-  is_primary: boolean;
-  services: string[];
-}
-
-export interface Caseworker {
-  firstName: string;
-  lastName: string;
-  idamId: string;
-  email: string;
-  location: Location;
-}
-
-export interface Judicialworker {
-  firstName: string;
-  lastName: string;
-  idamId: string;
-  email: string;
-  location: Location;
-}
-
-export interface CaseworkerApi {
-  first_name: string;
-  last_name: string;
-  id: string;
-  email_id: string;
-  base_location: LocationApi[];
-}
-
-export interface Action {
-  id: string;
-  title: string;
-}
