+++ src/work-allocation-2/containers/task-list-wrapper/task-list-wrapper.component.ts
@@ -2,7 +2,7 @@ import { ChangeDetectorRef, Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { AlertService, LoadingService } from '@hmcts/ccd-case-ui-toolkit';
 import { FeatureToggleService } from '@hmcts/rpx-xui-common-lib';
-import { Caseworker, Location } from 'api/workAllocation2/interfaces/task';
+import { Caseworker, Location } from 'api/workAllocation2/interfaces/common';
 import { Observable } from 'rxjs';
 import { SessionStorageService } from '../../../app/services';
 import { ListConstants } from '../../components/constants';
